fix: move rename home news events content
