+++ app/site-config/home/home-card-hero.tsx
